Pipe: refactor package structure (confignode)
