--- confignode/src/main/java/org/apache/iotdb/confignode/consensus/response/pipe/PipeSinkResp.java
@@ -1,48 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.apache.iotdb.confignode.consensus.response.pipe;
-
-import org.apache.iotdb.common.rpc.thrift.TSStatus;
-import org.apache.iotdb.commons.sync.pipesink.PipeSink;
-import org.apache.iotdb.consensus.common.DataSet;
-
-import java.util.List;
-
-public class PipeSinkResp implements DataSet {
-
-  TSStatus status;
-  List<PipeSink> pipeSinkList;
-
-  public TSStatus getStatus() {
-    return status;
-  }
-
-  public void setStatus(TSStatus status) {
-    this.status = status;
-  }
-
-  public List<PipeSink> getPipeSinkList() {
-    return pipeSinkList;
-  }
-
-  public void setPipeSinkList(List<PipeSink> pipeSinkList) {
-    this.pipeSinkList = pipeSinkList;
-  }
-}
