+++ confignode/src/main/java/org/apache/iotdb/confignode/procedure/impl/pipe/PipeTaskOperation.java
@@ -17,32 +17,14 @@
  * under the License.
  */
 
-package org.apache.iotdb.confignode.consensus.response.pipe;
+package org.apache.iotdb.confignode.procedure.impl.pipe;
 
-import org.apache.iotdb.common.rpc.thrift.TSStatus;
-import org.apache.iotdb.commons.sync.pipe.PipeInfo;
-import org.apache.iotdb.consensus.common.DataSet;
-
-import java.util.List;
-
-public class PipeResp implements DataSet {
-
-  TSStatus status;
-  List<PipeInfo> pipeInfoList;
-
-  public TSStatus getStatus() {
-    return status;
-  }
-
-  public void setStatus(TSStatus status) {
-    this.status = status;
-  }
-
-  public List<PipeInfo> getPipeInfoList() {
-    return pipeInfoList;
-  }
-
-  public void setPipeInfoList(List<PipeInfo> pipeInfoList) {
-    this.pipeInfoList = pipeInfoList;
-  }
+public enum PipeTaskOperation {
+  CREATE_PIPE,
+  START_PIPE,
+  STOP_PIPE,
+  DROP_PIPE,
+  HANDLE_LEADER_CHANGE,
+  SYNC_PIPE_META,
+  HANDLE_PIPE_META_CHANGE
 }
